Add document symbol provider for pester files
